requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.0.0; python_version >= '3.6'
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==4.2.4; python_version ~= '3.5'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9; python_version >= '3'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
commonmark==0.9.1
cryptography==36.0.1; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
decorator==4.4.2
flatbuffers==2.0
fonttools==4.28.5; python_version >= '3.7'
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==2.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
grpcio==1.43.0; python_version >= '3.6'
h5py==3.6.0; python_version >= '3.7'
humanfriendly==10.0
idna==3.3; python_version >= '3'
imageio-ffmpeg==0.4.5; python_version >= '3.4'
imageio==2.13.5; python_version >= '3.4'
importlib-metadata==4.10.0; python_version >= '3.7'
itsdangerous==2.0.1; python_version >= '3.6'
jeepney==0.7.1; sys_platform == 'linux'
jsonpickle==2.0.0
keras-preprocessing==1.1.2
keras==2.7.0
keyring==23.4.0; python_version >= '3.6'
kiwisolver==1.3.2; python_version >= '3.7'
libclang==12.0.0
markdown==3.3.6; python_version >= '3.6'
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.21.5; python_version < '3.11' and python_version >= '3.7'
oauthlib==3.1.1; python_version >= '3.6'
opencv-python==4.5.4.60
opt-einsum==3.3.0; python_version >= '3.5'
packaging==21.3; python_version >= '3.6'
pandas==1.3.5
pillow==8.4.0
proglog==0.1.9
protobuf==3.19.1; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pygments==2.10.0; python_version >= '3.5'
pyinaturalist==0.15.0
pyparsing==3.0.6; python_version >= '3.6'
pyqt5-plugins==5.15.4.2.2
pyqt5-qt5==5.15.2
pyqt5-sip==12.9.0; python_version >= '3.5'
pyqt5==5.15.4
pyrate-limiter==2.6.0; python_version < '4' and python_full_version >= '3.6.2'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-forge==18.6.0
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
qt5-applications==5.15.2.2.2; python_version >= '3.5'
qt5-tools==5.15.2.1.2; python_version >= '3.5'
requests-cache==0.7.5; python_version >= '3.6' and python_version < '4'
requests-oauthlib==1.3.0
requests-ratelimiter==0.2.1; python_version < '4' and python_full_version >= '3.6.2'
requests==2.26.0
rich==10.16.1; python_version < '4' and python_full_version >= '3.6.2'
rsa==4.8; python_version >= '3.6'
secretstorage==3.3.1; sys_platform == 'linux'
setuptools==60.0.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.7.0; python_version >= '3.6'
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1; python_version < '3.11' and python_version >= '3.7'
tensorflow==2.7.0
termcolor==1.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.0.1; python_version >= '3.6'
url-normalize==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.6.0; python_version >= '3.6'
